1st code refactoring
