+++ everytime/everytime/src/main/java/com/ceos19/everytime/domain/AboutCourse/AddedCourse.java
@@ -1,0 +1,26 @@
+package com.ceos19.everytime.domain.AboutCourse;
+
+import com.ceos19.everytime.domain.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Getter
+@Builder
+public class AddedCourse extends BaseTimeEntity {
+
+    @Id
+    @Column(name="added_course_id", nullable = false)
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    private Long addedCourseId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="timetable_id")
+    private Timetable timetable;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="course_id")
+    private Course course;
+}
