+++ everytime/everytime/src/main/java/com/ceos19/everytime/domain/AboutPost/Scrap.java
@@ -1,0 +1,27 @@
+package com.ceos19.everytime.domain.AboutPost;
+
+import com.ceos19.everytime.domain.AboutPost.Post;
+import com.ceos19.everytime.domain.AboutUser.User;
+import com.ceos19.everytime.domain.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Getter
+@Builder
+public class Scrap extends BaseTimeEntity {
+    @Id
+    @Column(name="scrap_id", nullable = false)
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    private Long scrapId;
+
+    @ManyToOne(fetch=FetchType.LAZY)
+    @JoinColumn(name="post_id")
+    private Post post;
+
+    @ManyToOne(fetch=FetchType.LAZY)
+    @JoinColumn(name="user_id")
+    private User user;
+}
