+++ everytime/everytime/src/main/java/com/ceos19/everytime/domain/AboutUser/Friend.java
@@ -1,0 +1,31 @@
+package com.ceos19.everytime.domain.AboutUser;
+
+
+import com.ceos19.everytime.domain.AboutUser.User;
+import com.ceos19.everytime.domain.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Getter
+@Builder
+public class Friend extends BaseTimeEntity {
+    @Id
+    @Column(name="friend_id", nullable = false)
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    private Long friendId;
+
+    @Column(name="is_accepted", nullable = false)
+    private boolean isAccepted=false;
+
+    @ManyToOne
+    @JoinColumn(name="request_user_id")
+    private User user1;
+
+    @ManyToOne
+    @JoinColumn(name="response_user_id")
+    private User user2;
+
+}
