+++ everytime/everytime/src/main/java/com/ceos19/everytime/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.ceos19.everytime.repository;
+
+import com.ceos19.everytime.domain.AboutUser.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUsername(String username);
+}
