--- everytime/src/main/java/com/ceos19/everytime/domain/AddedCourse.java
@@ -1,33 +1,0 @@
-package com.ceos19.everytime.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.sql.Timestamp;
-
-@Entity
-@Table(name="AddedCourse")
-@Getter
-@Setter
-public class AddedCourse {
-
-    @Id
-    @Column(name="added_course_id", nullable = false)
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long added_course_id;
-
-    @Column(name="created_at", nullable = false)
-    private Timestamp created_at;
-
-    @Column(name="updated_at", nullable = false)
-    private Timestamp updated_at;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="timetable_id")
-    private Timetable timetable;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="course_id")
-    private Course course;
-}
