--- everytime/src/main/java/com/ceos19/everytime/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.ceos19.everytime.repository;
-
-import com.ceos19.everytime.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByUsername(String username);
-}
